pyproject.toml
[tool.poetry]
name = "psqlomni"
version = "0.1.1"
description = "LLM-powered chat interface to your Postgres database - (psql powered with Natural Language)"
authors = ["emmakodes <emmamichael65@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
prompt-toolkit = "^3.0.43"
psycopg2-binary = "^2.9.9"
toml = "^0.10.2"
langchain = "^1.2.10"
langchain-community = "^0.4.1"
langchain-openai = "^1.1.9"
langgraph = "^1.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
psqlomni = 'psqlomni.__main__:main'
